+++ SwiftySCAD/Constants.swift
@@ -4,6 +4,6 @@
 
 import Foundation
 
-let epsilon: CGFloat = 0.001
-let epsilon_2: CGFloat = epsilon/2.0
-let big_epsilon: CGFloat = 1
+let epsilon: Float = 0.001
+let epsilon_2: Float = epsilon/2.0
+let big_epsilon: Float = 1
